requirement.txt
asgiref==3.3.1
Django==3.1.3
django-debug-toolbar==3.1.1
isort==5.6.4
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
